Add files via upload
Write an ANSI C program, called MusicLibrary, that allows the user to manage their music collection.  Each MusicLibrary has a name specified as an argument to the command line.  If none is given, use the default name of myMusic.  When the program is run, the named MusicLibrary is loaded from the file (name of file matches name of MusicLibrary) into memory.  If the file does not exist, the in-memory MusicLibrary will be empty.  When the program is exited, the MusicLibrary is saved back to the file from which it came. Put your entire program in one file named MusicLibrary.c. You will compile your program with the command gcc MusicLibrary.c -o MusicLibrary

A MusicLibrary is a list of songs - each of which has a title, artist, and year published.  Note we may add more fields in the future, but for now, just store those three values.  Songs are retrieved or deleted by using their title. You may limit each title and artist to 40 characters.

Your main program will be a loop that prints a